Simplify snake drawing
The original intention was to draw the front of the snake perpendicular
to its heading, like a bent line. This is kind of a lot of work so
instead we just blit a cell over the snake's current position and erase
a cell over the snake's last trail position. The visual effect
serendipitously looks like a chevron. Let's just keep it that way.